Renamed remote_management/hpe to remote_management/oneview
